backend/requirements.txt
click==8.1.3
Flask==2.1.2
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.1.1
python-dotenv==0.20.0
pytube==12.1.0
six==1.16.0
Werkzeug==2.1.2
